Issue #SB-23256 fix: testcase issue fix
