+++ src/app/client/src/app/modules/core/services/content-helper/content-helper.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ContentHelperService } from './content-helper.service';
+
+xdescribe('ContentHelperService', () => {
+  let service: ContentHelperService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ContentHelperService);
+  });
+
+  xit('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
