+++ src/app/client/src/app/modules/core/services/index.ts
@@ -25,4 +25,4 @@ export * from './enroll-contributor/enroll-contributor.service';
 export * from './textbook-list/textbook-list.service';
 export * from './registry/registry.service';
 export * from './notification/notification.service';
-export * from './contentHelper/contentHelper.service';
+export * from './content-helper/content-helper.service';
